requirements.txt
streamlit>=1.26.0
opencv-python>=4.8.0
tensorflow>=2.13.0
numpy>=1.24.0
vit-keras>=0.1.0
tensorflow-addons>=0.20.0
gdown>=4.7.1
pathlib>=1.0.1